qwerty
qwerty

+++ My project/Assets/Scripts/Entity/Interface/IDamageable.cs
@@ -2,5 +2,5 @@
 
 public interface IDamageable
 {
-    public void Damaged(int damage);
+    public void Damaged(float damage);
 }
